Tweak some wording.
The result is basically the same, but where Nervos has lock scripts
which check signatures, Cardano has dedicated support for signature
checking; no validator script is needed if all you want to do is
check a signature. So this patch rewords things to account for that,
though it fundamentally doesn't change anything.